+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.4.1
 
 * Added load statements for all cc_binary, cc_library, cc_test functions in all bazel files.
+* Added `mbo::strings::ConsumePrefix` and `mbo::strings::ConsumeSuffix`.
 
 # 0.4.0